requirements.txt
streamlit==1.44.1
pandas==2.2.3
numpy==2.2.5
python-dotenv==1.1.0
requests==2.31.0